Use precompiled v release (#16)
* Use pre-compiled release of V

Turns out my solution for specifying the version of V to use in the
runner in PR #12 was flawed - the build process quietly ignores most of
the contents of the repo and pulls the latest pre-transpiled version of
the compiler from vlang/vc and builds that instead. So pinning the
build repo to a particular release means it will eventually pull an
incompatible version of vc, as has now happened.

We could bite the bullet and just build the latest everytime, but this
seems like a recipe for disaster given that the runner could behave
differently on every run. Instead, we take advantage of official
pre-compiled releases of V now being available. Using them required
some changes to the image, but we end up in a good place that
preserves the test integrity of the runner and adds a predictable,
mainstream version of V.

* Remove unnecessary, non-portable "[[" keyword.

The "[[" keyword requires specific shell support. This creates an
an unnecessary requirement and currently fails when running the
script in Docker. The "[" command is sufficient.

* Fix test results

---------

Co-authored-by: Erik Schierboom <erik_schierboom@hotmail.com>